python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Test and demo for Azure AI Foundry Python SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai==2.30.0",
    "azure-ai-projects==2.0.1",
    "python-dotenv==1.2.2",
    "azure-identity==1.25.3"
]
